Title: 합과 차의 나눗셈, Time: 112ms, Memory: 24MB, Status: Wrong Answer - CodeTree

+++ 240109/합과 차의 나눗셈/divide-of-sum-and-sub.py
@@ -1,4 +1,4 @@
-a, b = map(int, input().split())
+a, b = map(float, input().split())
 
 m = (a+b) / (a-b)
 print(round(m,2))
